Refactor OAUTH methods into DashService
