+++ dspace/modules/api/src/main/java/org/datadryad/api/DashService.java
@@ -1,0 +1,155 @@
+package org.datadryad.api;
+
+/** 
+    Utilities to facilitate communication with the DASH-based version of Dryad.
+ **/
+
+import java.io.BufferedReader;
+import java.io.InputStreamReader;
+import java.io.IOException;
+import java.io.OutputStreamWriter;
+import java.io.PrintStream;
+import java.io.StringWriter;
+import java.net.HttpURLConnection;
+import java.net.URL;
+import java.util.Base64;
+import java.util.regex.Matcher;
+import java.util.regex.Pattern;
+import org.apache.log4j.Logger;
+import org.dspace.core.ConfigurationManager;
+
+
+
+
+public class DashService {
+
+    private static final Logger log = Logger.getLogger(DashService.class);
+    private String dashServer = "";
+    private String oauthToken = "";
+
+    public DashService() {
+        // init oauth connection with DASH
+        dashServer = ConfigurationManager.getProperty("dash.server");
+        String dashAppID = ConfigurationManager.getProperty("dash.application.id");
+        String dashAppSecret = ConfigurationManager.getProperty("dash.application.secret");
+        
+        oauthToken = getOAUTHtoken(dashServer, dashAppID, dashAppSecret);
+    }
+
+    
+    private String getOAUTHtoken(String dashServer, String dashAppID, String dashAppSecret) {
+        
+        String url = dashServer + "/oauth/token";
+        String auth = dashAppID + ":" + dashAppSecret;
+        String authentication = Base64.getEncoder().encodeToString(auth.getBytes());
+        Pattern tokenPattern = Pattern.compile(".*\"access_token\"\\s*:\\s*\"([^\"]+)\".*");
+        String token = "";
+
+        try {
+            URL urlObj = new URL(url);
+            HttpURLConnection con = (HttpURLConnection) urlObj.openConnection();
+            con.setRequestMethod("POST");
+            con.setDoOutput(true);
+            con.setRequestProperty("Authorization", "Basic " + authentication);
+            con.setRequestProperty("Content-Type", "application/x-www-form-urlencoded;charset=UTF-8");
+            con.setRequestProperty("Accept", "application/json");
+            
+            PrintStream os = new PrintStream(con.getOutputStream());
+            os.print("grant_type=client_credentials");
+            os.close();
+            BufferedReader reader = new BufferedReader(new InputStreamReader(con.getInputStream()));
+            String line = null;
+            StringWriter out = new StringWriter(con.getContentLength() > 0 ? con.getContentLength() : 2048);
+            while ((line = reader.readLine()) != null) {
+                out.append(line);
+            }
+            String response = out.toString();
+            Matcher matcher = tokenPattern.matcher(response);
+            if (matcher.matches() && matcher.groupCount() > 0) {
+                token = matcher.group(1);
+            }
+
+            log.info("got OAuth token " + token);
+
+        } catch (Exception e) {
+            log.fatal("Unable to obtain OAuth token", e);
+        }
+
+        return token;
+    }
+
+    public String getDashJSON(String doi) {
+        return "getDashJSON NOT IMPLEMENTED YET";
+    }
+    
+    public String postDataPackage(DryadDataPackage dataPackage) {
+        String dashJSON = dataPackage.getDashJSON();
+        String responseCode = "POST not completed";
+        log.debug("Got JSON object: " + dashJSON);
+            
+        BufferedReader reader = null;
+
+        //TODO: replace this pattern matching with real JSON parsing 
+        Pattern datasetIDPattern = Pattern.compile("(/api/datasets/.+?)\"},"); 
+        
+        try {
+            URL url = new URL(dashServer + "/api/datasets");
+            HttpURLConnection connection = (HttpURLConnection) url.openConnection();
+            connection.setRequestProperty("Content-Type", "application/json; charset=UTF-8");
+            connection.setRequestProperty("Accept", "application/json");
+            connection.setRequestProperty("Authorization", "Bearer " + oauthToken);
+            connection.setDoOutput(true);
+            connection.setRequestMethod("POST");
+
+            OutputStreamWriter wr = new OutputStreamWriter(connection.getOutputStream());
+            wr.write(dashJSON);
+            wr.close();
+
+            reader = new BufferedReader(new InputStreamReader(connection.getInputStream()));
+            String line = null;
+            StringWriter out = new StringWriter(connection.getContentLength() > 0 ? connection.getContentLength() : 2048);
+            while ((line = reader.readLine()) != null) {
+                out.append(line);
+            }
+
+            responseCode = connection.getResponseCode() + " " + connection.getResponseMessage();
+            String response = out.toString();
+            String datasetID = "";
+            Matcher matcher = datasetIDPattern.matcher(response);
+            if (matcher.matches() && matcher.groupCount() > 0) {
+                datasetID = matcher.group(1);
+            }
+
+            log.info("got dataset ID " + datasetID);
+            log.info("result object " + response);
+        } catch (Exception e) {
+            log.fatal("Unable to send item to DASH", e);
+        }
+
+        return responseCode;
+    }
+    
+    /**
+     * Have to test this on dev since it's also IP restricted.
+     *
+     * @param args
+     */
+    public static void main(String[] args) throws IOException {
+        
+        String usage = "\n\nUsage: \n" +
+            "\tlookup a specific item: class doi\n";
+        DashService service = new DashService();
+
+        log.debug("========== Starting DOI command-line service ===========");
+
+        // LOOKUP: args[0]=DOI
+        if (args.length == 1) {
+            String doiID = args[0];
+            System.out.println(service.getDashJSON(doiID));
+        } else {
+            System.out.println(usage);
+        }
+    }
+
+
+}
